requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
greenlet==1.1.2
idna==3.3
kombu==5.2.2
prompt-toolkit==3.0.22
psycopg2==2.9.2
Pygments==2.10.0
pymemcache==3.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
rich==9.13.0
six==1.16.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
tabulate==0.8.9
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
